requirements.txt
cupy_cuda115==10.6.0
numpy==1.24.4
opencv_python==4.9.0.80
scipy==1.12.0
torch==2.2.0+cu121
